structurizr-dsl-to-mermaid.yml
This workflow aims to automate the process of exporting Structurizr diagrams from DSL files and updating Mermaid files. It checks for existing Mermaid files and updates them instead of creating new ones when the DSL files are modified. The Mermaid files are stored in a ```mermaid/``` directory for further use, such as manual embedding in documentation.

1. Checks out the repository
2. Installs Structurizr CLI
3. Process DSL files
    - Iterates through Structurizr DSL files in the repository
    - Extracts the user name and base name from the file path
    - Checks if the corresponding Mermaid file already exists
    - If the Mermaid file exists, updates it using structurizr-cli export with the --force option.
    - If the Mermaid file doesn't exist, creates it using structurizr-cli export.